refactor: Move common overlay input logic to Uno.UI
